Handle potential arith o/flow in Storage_Offset comparisons
SPARK Discovery GPL 2017 has stricter handling of Storage_Offset than
prior versions of SPARK. It no longer assumes that Storage_Array lengths
will always fit into a Long_Long_Integer, so Storage_Array'Length can
no longer be referred to, and Storage_Offset calculations are liable to
overflows.

Several subroutine preconditions have to be re-written to include checks
that the Storage_Arrays being passed are short enough for their lengths
to fit in a Long_Long_Integer. In practice this will never be a problem.
Ghost functions are used to avoid the preconditions becoming unreadable.

Thanks to Claire Dross from AdaCore for the explanation and suggestions.